1440 (#1)
* Only leader can set resharding flag (#1472)

* updates
