+++ elasticjob-error-handler/elasticjob-error-handler-email/pom.xml
@@ -31,7 +31,7 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.shardingsphere.elasticjob</groupId>
-            <artifactId>elasticjob-error-handler-type</artifactId>
+            <artifactId>elasticjob-error-handler-general</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
