+++ elasticjob-error-handler/elasticjob-error-handler-email/src/main/resources/error-handler-email.yaml
@@ -15,4 +15,13 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.error.handler.fixture.LogJobErrorHandler
+email:
+  host: xxx
+  port: xxx
+  username: username
+  password: password
+  protocol: smtp
+  from: xxx
+  to: xxx
+  cc: xxx
+
