+++ elasticjob-error-handler/elasticjob-error-handler-email/src/test/resources/error-handler-email.yaml
@@ -15,6 +15,12 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.error.handler.general.LogJobErrorHandler
-org.apache.shardingsphere.elasticjob.error.handler.general.IgnoreJobErrorHandler
-org.apache.shardingsphere.elasticjob.error.handler.general.ThrowJobErrorHandler
+email:
+  host: test.mail.com
+  port: 123
+  username: username
+  password: password
+  protocol: smtp
+  from: testmail@qiyi.com
+  to: xxx1@ejob.com
+  cc: xxx2@ejob.com
