+++ elasticjob-error-handler/elasticjob-error-handler-general/src/main/java/org/apache/shardingsphere/elasticjob/error/handler/impl/ThrowJobErrorHandler.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,39 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.error.handler.fixture;
+package org.apache.shardingsphere.elasticjob.error.handler.impl;
 
-import lombok.Getter;
-import lombok.Setter;
-
-import java.nio.charset.Charset;
+import org.apache.shardingsphere.elasticjob.error.handler.JobErrorHandler;
+import org.apache.shardingsphere.elasticjob.infra.exception.JobSystemException;
 
 /**
- * Job configuration POJO.
+ * Job error handler for throw exception.
  */
-@Getter
-@Setter
-public final class EmailConfiguration {
-
-    private String host;
-
-    private Integer port;
-
-    private String username;
-
-    private String password;
-
-    private String protocol = "smtp";
-
-    private Charset encoding = Charset.forName("UTF-8");
-
-    private String subject = "Ejob error message";
-
-    private String from;
-
-    private String to;
-
-    private String cc;
-
-    private String bcc;
+public final class ThrowJobErrorHandler implements JobErrorHandler {
+    
+    @Override
+    public void handleException(final String jobName, final Throwable cause) {
+        throw new JobSystemException(cause);
+    }
+    
+    @Override
+    public String getType() {
+        return "THROW";
+    }
 }
