+++ elasticjob-error-handler/elasticjob-error-handler-general/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.error.handler.JobErrorHandler
@@ -15,6 +15,6 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.error.handler.general.LogJobErrorHandler
-org.apache.shardingsphere.elasticjob.error.handler.general.IgnoreJobErrorHandler
-org.apache.shardingsphere.elasticjob.error.handler.general.ThrowJobErrorHandler
+org.apache.shardingsphere.elasticjob.error.handler.impl.LogJobErrorHandler
+org.apache.shardingsphere.elasticjob.error.handler.impl.IgnoreJobErrorHandler
+org.apache.shardingsphere.elasticjob.error.handler.impl.ThrowJobErrorHandler
