+++ elasticjob-error-handler/elasticjob-error-handler-general/src/test/java/org/apache/shardingsphere/elasticjob/error/handler/impl/IgnoreJobErrorHandlerTest.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *  
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,24 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.error.handler.general;
+package org.apache.shardingsphere.elasticjob.error.handler.impl;
 
-import lombok.extern.slf4j.Slf4j;
-import org.apache.shardingsphere.elasticjob.error.handler.JobErrorHandler;
+import org.junit.Test;
 
-/**
- * Job error handler for log error message.
- */
-@Slf4j
-public final class LogJobErrorHandler implements JobErrorHandler {
-    
-    @Override
-    public void handleException(final String jobName, final Throwable cause) {
-        log.error(String.format("Job '%s' exception occur in job processing", jobName), cause);
-    }
-    
-    @Override
-    public String getType() {
-        return "LOG";
+public final class IgnoreJobErrorHandlerTest {
+
+    @Test
+    public void assertHandleException() {
+        new IgnoreJobErrorHandler().handleException("test_job", new RuntimeException("test"));
     }
 }
