+++ elasticjob-error-handler/pom.xml
@@ -33,6 +33,5 @@
         <module>elasticjob-error-handler-general</module>
         <module>elasticjob-error-handler-email</module>
         <module>elasticjob-error-handler-sms</module>
-        <module>elasticjob-error-handler-type</module>
     </modules>
 </project>
