+++ elasticjob-lite/elasticjob-lite-core/src/main/java/org/apache/shardingsphere/elasticjob/lite/internal/sharding/ShardingService.java
@@ -80,6 +80,9 @@ public final class ShardingService {
      * Set resharding flag.
      */
     public void setReshardingFlag() {
+        if (!leaderService.isLeaderUntilBlock()) {
+            return;
+        }
         jobNodeStorage.createJobNodeIfNeeded(ShardingNode.NECESSARY);
     }
