+++ elasticjob-lite/elasticjob-lite-core/src/test/java/org/apache/shardingsphere/elasticjob/lite/internal/reconcile/ReconcileServiceTest.java
@@ -20,7 +20,6 @@ package org.apache.shardingsphere.elasticjob.lite.internal.reconcile;
 import org.apache.shardingsphere.elasticjob.api.JobConfiguration;
 import org.apache.shardingsphere.elasticjob.infra.handler.sharding.JobInstance;
 import org.apache.shardingsphere.elasticjob.lite.internal.config.ConfigurationService;
-import org.apache.shardingsphere.elasticjob.lite.internal.election.LeaderService;
 import org.apache.shardingsphere.elasticjob.lite.internal.schedule.JobRegistry;
 import org.apache.shardingsphere.elasticjob.lite.internal.sharding.ShardingService;
 import org.apache.shardingsphere.elasticjob.lite.util.ReflectionUtils;
@@ -42,9 +41,6 @@ public final class ReconcileServiceTest {
     @Mock
     private ShardingService shardingService;
     
-    @Mock
-    private LeaderService leaderService;
-    
     private ReconcileService reconcileService;
     
     @Before
@@ -54,7 +50,6 @@ public final class ReconcileServiceTest {
         ReflectionUtils.setFieldValue(reconcileService, "lastReconcileTime", 1L);
         ReflectionUtils.setFieldValue(reconcileService, "configService", configService);
         ReflectionUtils.setFieldValue(reconcileService, "shardingService", shardingService);
-        ReflectionUtils.setFieldValue(reconcileService, "leaderService", leaderService);
     }
     
     @Test
@@ -62,11 +57,9 @@ public final class ReconcileServiceTest {
         when(configService.load(true)).thenReturn(JobConfiguration.newBuilder("test_job", 3).cron("0/1 * * * * ?").reconcileIntervalMinutes(1).build());
         when(shardingService.isNeedSharding()).thenReturn(false);
         when(shardingService.hasShardingInfoInOfflineServers()).thenReturn(true);
-        when(leaderService.isLeaderUntilBlock()).thenReturn(true);
         reconcileService.runOneIteration();
         verify(shardingService).isNeedSharding();
         verify(shardingService).hasShardingInfoInOfflineServers();
         verify(shardingService).setReshardingFlag();
-        verify(leaderService).isLeaderUntilBlock();
     }
 }
